Move bits and pieces to correct package
